# app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/system/directories": {
            "get": {
                "description": "\"获取所有配置目录\"",
                "tags": [
                    "system"
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/i18n.CustomError"
                        }
                    }
                }
            }
        },
        "/system/proxy": {
            "get": {
                "description": "\"代理请求\"",
                "tags": [
                    "system"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "请求url",
                        "name": "url",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/i18n.CustomError"
                        }
                    }
                }
            }
        },
        "/vscode/copy": {
            "post": {
                "description": "将旧URI重命名为新URI。如果overwrite为true，则覆盖同名文件",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "vscode"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "source",
                        "name": "source",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "destination",
                        "name": "destination",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "是否覆盖同名文件",
                        "name": "overwrite",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "复制成功"
                    },
                    "400": {
                        "description": "复制失败"
                    }
                }
            }
        },
        "/vscode/createDirectory": {
            "post": {
                "description": "创建一个新目录",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "vscode"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "URI",
                        "name": "uri",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功"
                    },
                    "400": {
                        "description": "创建失败"
                    }
                }
            }
        },
        "/vscode/delete": {
            "delete": {
                "description": "删除指定URI的文件或目录。如果recursive为true，则删除所有子目录和文件",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "vscode"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "URI",
                        "name": "uri",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "是否删除所有子目录和文件",
                        "name": "recursive",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功"
                    },
                    "400": {
                        "description": "删除失败"
                    }
                }
            }
        },
        "/vscode/readDirectory": {
            "get": {
                "description": "返回指定URI下的所有文件和目录的元数据",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "vscode"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "URI",
                        "name": "uri",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "忽略不存在报错",
                        "name": "ignoreNotExist",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/vscode.FileStat"
                            }
                        }
                    },
                    "400": {
                        "description": "查询失败"
                    }
                }
            }
        },
        "/vscode/readFile": {
            "get": {
                "description": "返回指定URI的文件内容",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "vscode"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "URI",
                        "name": "uri",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "查询失败"
                    }
                }
            }
        },
        "/vscode/rename": {
            "put": {
                "description": "将旧URI重命名为新URI。如果overwrite为true，则覆盖同名文件",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "vscode"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "oldURI",
                        "name": "oldURI",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "newURI",
                        "name": "newURI",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "是否覆盖同名文件",
                        "name": "overwrite",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功"
                    },
                    "400": {
                        "description": "删除失败"
                    }
                }
            }
        },
        "/vscode/stat": {
            "get": {
                "description": "返回指定URI的文件元数据",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "vscode"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "URI",
                        "name": "uri",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "$ref": "#/definitions/vscode.FileStat"
                        }
                    },
                    "400": {
                        "description": "查询失败"
                    }
                }
            }
        },
        "/vscode/writeFile": {
            "post": {
                "description": "将content写入指定URI的文件中。如果create为true，则在文件不存在的情况下创建文件；如果overwrite",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "vscode"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "URI",
                        "name": "uri",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "URI",
                        "name": "content",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "create",
                        "name": "create",
                        "in": "formData"
                    },
                    {
                        "type": "boolean",
                        "description": "overwrite",
                        "name": "overwrite",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "写入成功"
                    },
                    "400": {
                        "description": "写入失败"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "i18n.CustomError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "错误码",
                    "allOf": [
                        {
                            "$ref": "#/definitions/i18n.Errcode"
                        }
                    ]
                },
                "message": {
                    "description": "错误消息",
                    "type": "string"
                },
                "mode": {
                    "description": "业务模块",
                    "type": "string"
                }
            }
        },
        "i18n.Errcode": {
            "type": "integer",
            "enum": [
                10101,
                10201,
                10202,
                10203,
                10204,
                10205,
                10206,
                10207,
                10301,
                10302,
                10303,
                10304,
                10305,
                10401,
                10402,
                10403,
                10404,
                10405,
                10406,
                10407,
                10501,
                10502,
                10503,
                10504,
                10505,
                10506,
                10507,
                10508,
                10509,
                10510,
                10511,
                10512,
                10513,
                10514,
                10515,
                10516,
                10517,
                10518,
                10519,
                10520,
                10601,
                10602,
                10603,
                10604,
                10605,
                10606,
                20101,
                10102,
                20201,
                20202,
                20203,
                20204,
                20205,
                20206,
                20207,
                20208,
                20209,
                20210,
                20211,
                20301,
                20302,
                20303,
                20304,
                20305,
                20306,
                20307,
                20401,
                20402,
                20403,
                20404,
                20405,
                20406,
                20407,
                20408,
                20409,
                20501,
                20502,
                20601,
                20701
            ],
            "x-enum-varnames": [
                "ServerError",
                "ParamIllegalError",
                "ParamBindError",
                "StructParamEmtpyError",
                "BodyParamEmptyError",
                "PathParamEmptyError",
                "QueryParamEmptyError",
                "FormParamEmptyError",
                "RequestResubmitError",
                "RequestSignatureError",
                "RequestReadBodyError",
                "RequestEmptyBodyError",
                "RequestProxyError",
                "FileReadError",
                "FileWriteError",
                "FileZipError",
                "FileUnZipError",
                "FileZipAmountZeroError",
                "FileContentEmptyError",
                "DirectoryReadError",
                "LoaderFileReadError",
                "LoaderFileUnmarshalError",
                "LoaderDataExistEditorError",
                "LoaderDataExistError",
                "LoaderDataNotExistError",
                "LoaderLockNotFoundError",
                "LoaderWatcherNotSupport",
                "LoaderNoneModifiedError",
                "LoaderRWNotSupportError",
                "LoaderNameEmptyError",
                "LoaderBasenameEmptyErr",
                "LoaderRootOrExtensionEmptyErr",
                "LoaderMultipleOnlyError",
                "LoaderEmbedNotAllowModifyErr",
                "LoaderRemoveKeyNotFoundError",
                "LoaderRenameKeyNotFoundError",
                "LoaderRenameTargetExistError",
                "LoaderRenameNotAllowMultipleError",
                "LoaderWriteableRelyModelRequiredError",
                "LoaderDataFilepathError",
                "VscodeOnlyDirectoriesCanWatchError",
                "VscodeDirectoryExistError",
                "VscodeFileExistError",
                "VscodeFileNotExistError",
                "VscodeSourceNotDirectoryError",
                "VscodeTargetDirectoryExistError",
                "EngineCreateConfigError",
                "EngineRestartError",
                "DataInsertError",
                "DataDeleteError",
                "DataUpdateError",
                "DataSelectError",
                "DataCopyError",
                "DataRenameError",
                "DataBatchInsertError",
                "DataBatchDeleteError",
                "DataBatchUpdateError",
                "DataEmptyListError",
                "DataNotExistsError",
                "DatasourceConnectionError",
                "DatasourceKindNotSupportedError",
                "DatasourceDisabledError",
                "DatasourceDatabaseUrlEmptyError",
                "DatabaseOasVersionError",
                "PrismaQueryError",
                "PrismaMigrateError",
                "StoragePingError",
                "StorageDisabledError",
                "StorageMkdirError",
                "StorageTouchError",
                "StorageRemoveError",
                "StorageRenameError",
                "StorageListError",
                "StorageDetailError",
                "StorageDownloadError",
                "OperationRoleHasBindError",
                "OperationRbacTypeError",
                "SettingServerUrlEmptyError",
                "SdkAlreadyUpToDateError"
            ]
        },
...
        "vscode.FileStat": {
            "type": "object",
            "properties": {
                "ctime": {
                    "type": "integer"
                },
                "mtime": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "type": {
                    "$ref": "#/definitions/vscode.FileType"
                }
            }
        },
        "vscode.FileType": {
            "type": "integer",
            "enum": [
                1,
                2,
                64
            ],
            "x-enum-varnames": [
                "File",
                "Directory",
                "SymbolicLink"
            ]
        }
    }
}